feat: new post script; refactor main logic into run function
