+++ src/post.ts
@@ -1,0 +1,4 @@
+import * as toolbox from '@wealthsimple/actions-toolbox';
+import { run } from './run';
+
+run(toolbox.post.run);
